@@ -2,22 +2,22 @@ import 'package:flutter/material.dart';
 
 class Badge extends StatelessWidget {
   const Badge({
-    Key key,
-    @required this.child,
-    @required this.value,
+    Key? key,
+    required this.child,
+    required this.value,
     this.color,
   }) : super(key: key);
 
-  final Widget child;
+  final Widget? child;
   final String value;
-  final Color color;
+  final Color? color;
 
   @override
   Widget build(BuildContext context) {
     return Stack(
       alignment: Alignment.center,
       children: [
-        child,
+        child!,
         Positioned(
           right: 8,
           top: 8,
