@@ -50,7 +50,7 @@ class Products with ChangeNotifier {
     return _items.where((prodItem) => prodItem.isFavorite).toList();
   }
 
-  Product findById(String id) {
+  Product findById(String? id) {
     return _items.firstWhere((prod) => prod.id == id);
   }
 
