@@ -8,7 +8,7 @@ class ProductDetailScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final productId =
-        ModalRoute.of(context).settings.arguments as String; // is the id!
+        ModalRoute.of(context)!.settings.arguments as String?; // is the id!
     final loadedProduct = Provider.of<Products>(
       context,
       listen: false,
